feat(skore/login): Assert optional dependencies are installed on the fly
