+++ skore/tests/unit/project/test_login.py
@@ -1,6 +1,7 @@
 from importlib.metadata import EntryPoint, EntryPoints
 from re import escape
 from unittest.mock import Mock
+from uuid import uuid4
 
 from pytest import fixture, raises
 
@@ -19,6 +20,7 @@ class FakeEntryPoint(EntryPoint):
 
 
 def test_login_local(monkeypatch, FakeLogin):
+    monkeypatch.setattr("skore._project.dependencies.requires", lambda _: [])
     monkeypatch.setattr(
         "skore._project.plugin.entry_points",
         lambda **kwargs: EntryPoints(
@@ -39,7 +41,28 @@ def test_login_local(monkeypatch, FakeLogin):
     assert not FakeLogin.called
 
 
+def test_login_local_missing_optional_dependency(monkeypatch):
+    fake_library_name = uuid4().hex
+
+    def fake_requires(name):
+        assert name == "skore"
+        return [f'{fake_library_name} ; extra == "local"']
+
+    monkeypatch.setattr("skore._project.dependencies.requires", fake_requires)
+
+    from skore import login
+
+    with raises(
+        ImportError,
+        match=escape(
+            f"Missing `{fake_library_name}` library: please install `skore[local]`."
+        ),
+    ):
+        login(mode="local", arg1=1, arg2=2)
+
+
 def test_login_hub(monkeypatch, FakeLogin):
+    monkeypatch.setattr("skore._project.dependencies.requires", lambda _: [])
     monkeypatch.setattr(
         "skore._project.plugin.entry_points",
         lambda **kwargs: EntryPoints(
@@ -62,7 +85,8 @@ def test_login_hub(monkeypatch, FakeLogin):
     assert FakeLogin.call_args.kwargs == {"arg1": 1, "arg2": 2}
 
 
-def test_login_hub_without_plugin(monkeypatch):
+def test_login_hub_unknown_plugin(monkeypatch):
+    monkeypatch.setattr("skore._project.dependencies.requires", lambda _: [])
     monkeypatch.setattr(
         "skore._project.plugin.entry_points", lambda **kwargs: EntryPoints([])
     )
